pyproject.toml
[tool.poetry]
name = "product-scraper-poc"
version = "0.1.0"
description = ""
authors = ["Tejus Ravi <tejusravidev@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
hydra-core = "^1.3.2"
loguru = "^0.7.2"
selenium = "^4.24.0"
aiohttp = "^3.10.5"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
